requirements.txt
# mlflow
mlflow==2.12.1
mlflow[genai]==2.12.1
psycopg2==2.9.9
boto3==1.34.94

python-dotenv==1.0.1

pandas==2.2.2
hyperopt==0.2.7
keras==3.3.3

torch==2.3.0
torchvision==0.18.0

langchain==0.1.10